pyproject.toml
[project]
name = "sports-card-analyzer"
version = "0.1.0"
description = "A Streamlit application for analyzing sports cards"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "plotly>=5.18.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv==1.0.1",
    "xlsxwriter>=3.2.0",
    "Pillow>=10.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.streamlit]
theme.base = "light"
theme.primaryColor = "#FF4B4B"
theme.backgroundColor = "#FFFFFF"
theme.secondaryBackgroundColor = "#F0F2F6"
theme.textColor = "#262730"
theme.font = "sans serif" 